requirements.txt
dash-auth
dash-bootstrap-components
django-cors-headers
django-extensions
djangorestframework
dtreeviz
flask-simplelogin
jupyter-dash
lightgbm
oyaml
pandasai
pip-chill
shap
sqlglotrs
waitress
xgboost
ydata-profiling==4.5.1
numba==0.57.1
whitenoise
